Structural improvements
-Moved update presence to util.js
-Changed "Playing x in KAG" to "Watching x in KAG"
-Moved updateServerList to serverlist.js
