+++ modules/util.js
@@ -1,4 +1,3 @@
-const Discord = require("discord.js");
 const XMLHttpRequest = require("xmlhttprequest").XMLHttpRequest;
 const config = require("../config.json");
 
@@ -7,7 +6,6 @@ var client;
 exports.init = (cl) => {
 	client = cl;
 }
-
 
 exports.getRole = (role) => {
 	if (!role) return null;
@@ -18,7 +16,7 @@ exports.getRole = (role) => {
 		role = client.guilds.get(config.guild).roles.find(x => x.name.toLowerCase() === role.toLowerCase());
 	}
 	return role || null;
-},
+}
 
 exports.getUser = (user) => {
 	if (!user) return null;
@@ -35,23 +33,23 @@ exports.getUser = (user) => {
 			|| guildMembers.find(x => (x.nickname || x.user.username).toLowerCase().includes(user.toLowerCase()));
 	}
 	return user || null;
-},
+}
 
 exports.getChannel = (channel) => {
 	if (typeof channel === "object") return channel;
 	return client.guilds.get(config.guild).channels.get(channel) || null;
-},
+}
 
 exports.userHasRole = (user, role) => {
 	user = this.getUser(user);
 	role = this.getRole(role);
 	if (!user || !role) return false;
 	return client.guilds.get(config.guild).members.get(user.id).roles.has(role.id);
-},
+}
 
 exports.plural = (val, text, suffix = "s") => {
 	return val === 1 ? text : text + suffix;
-},
+}
 
 exports.XMLHttpRequest = (callback, url) => {
 	let xhttp = new XMLHttpRequest();
@@ -66,7 +64,7 @@ exports.XMLHttpRequest = (callback, url) => {
 	}
 	xhttp.open("GET", url, true);
 	xhttp.send();
-},
+}
 
 exports.sendMessage = (channel, text, delete_message = 0) => {
     channel = this.getChannel(channel);
@@ -78,7 +76,7 @@ exports.sendMessage = (channel, text, delete_message = 0) => {
 	}).catch(err => {
 		console.log(`ERROR: Couldn't send message in #${message.channel.name} - ${err}`);
 	});
-},
+}
 
 exports.editMessage = (message, text, delete_message = 0) => {
 	message.edit(text).then(message => {
@@ -88,7 +86,7 @@ exports.editMessage = (message, text, delete_message = 0) => {
 	}).catch(err => {
 		console.log(`ERROR: Couldn't edit message in #${message.channel.name} - ${err}`);
 	});
-},
+}
 
 exports.deleteMessage = (message, delete_message = 0) => {
 	if (delete_message) {
@@ -102,7 +100,7 @@ exports.deleteMessage = (message, delete_message = 0) => {
 			console.log(`ERROR: Couldn't delete message in #${message.channel.name} - ${err}`);
 		});
 	}
-},
+}
 
 exports.fetchMessage = (callback, cfg_group) => {
 	channel = this.getChannel(cfg_group.channel);
@@ -115,7 +113,7 @@ exports.fetchMessage = (callback, cfg_group) => {
 	}).catch(() => {
 		console.log(`ERROR: Couldn't fetch message from #${channel.name}`);
 	});
-},
+}
 
 exports.sortServers = (servers) => {
 	if (servers.hasOwnProperty("serverList")) {
@@ -143,75 +141,23 @@ exports.sortServers = (servers) => {
 	}
 	
     return servers;
-},
+}
 
 exports.isMod = (user) => {
 	return config.mod_roles.some(role => {
 		return this.userHasRole(user, role);
 	});
-},
+}
 
-exports.updateServerList = (message, servers) => {
-	if (!message) return;
-	
-    let total_servers = servers.length;
-	let total_players = servers.reduce((t, x) => t + x.currentPlayers, 0);
+exports.updatePresence = () => {
+	this.XMLHttpRequest(servers => {
+		servers = this.sortServers(servers);
+		let total_players = servers.reduce((t, x) => t + x.currentPlayers, 0);
+		client.user.setActivity(`${total_players} in KAG | ${config.prefix}help`, { type: 'WATCHING' });
+	}, 'https://api.kag2d.com/v1/game/thd/kag/servers?filters=[{"field":"current","op":"eq","value":"true"},{"field":"connectable","op":"eq","value":true},{"field":"currentPlayers","op":"gt","value":"0"}]');
 
-	const embed = new Discord.RichEmbed()
-		.setTitle(":crossed_swords: KAG Server List :bow_and_arrow:")
-		.setColor(config.server_list.embed_colour)
-		.setDescription(`${total_servers} ${this.plural(total_servers, "server")} with ${total_players} ${this.plural(total_players, "player")}\n​`)
-		.setFooter("Last updated")
-		.setTimestamp();
-	
-	let count = 0;
-	for (let server of servers) {
-		//get flag for each server
-		this.XMLHttpRequest(data => {
-			count++;
-			if (!data) return;
-			server.region = data.country;
-
-			//add servers to embed once we have all flags
-			if (count === servers.length) {
-				for (let server of servers) {
-
-					let description = server.description;
-					if (server.usingMods) {
-						description = description.replace(/(\n\n|\n$)[^\n]*$/, "");
-					}
-					let spectators = server.spectatorPlayers ? ` (${server.spectatorPlayers} ${this.plural(server.spectatorPlayers, "spectator")})` : "";
-
-					//escape underscores so they dont italicise the text
-					let text = [
-						`**Description:** ${description.length ? description.replace(/_/g, "\\_") : "*no description*"}`,
-						`**Address:** ${server.password ? "*locked server*" : `<kag://${server.IPv4Address}:${server.port}>`}`,
-						`**Gamemode:** ${server.gameMode.replace(/_/g, "\\_")}`,
-						`**Players:** ${server.currentPlayers}/${server.maxPlayers}${spectators}`,
-						server.playerList.join(" ").replace(/_/g, "\\_"),
-						"​" //zero-width character for spacing
-					].sort(Boolean).join("\n");
-
-					//truncate text if too long
-					let ellipsis = "…";
-					if (text.length > 1024 - ellipsis.length)
-					{
-						//trim the string to the maximum length
-						text = text.substr(0, 1024 - ellipsis.length);
-						//re-trim if we are in the middle of a word
-						text = text.substr(0, Math.min(text.length, text.lastIndexOf(" ") + 1)) + ellipsis;
-					}
-
-					embed.addField(`:flag_${server.region.toLowerCase()}: ${server.name}`, text);
-				}
-				this.editMessage(message, embed);
-
-				if (config.server_list.special_channel_name) {
-					message.channel.setName(`${total_servers}-servers_${total_players}-players`).catch(() => {
-						console.log(`ERROR: Couldn't change name of #${message.channel.name}`);
-					});
-				}
-			}
-		}, `https://get.geojs.io/v1/ip/country/${server.IPv4Address}.json`);
-	}
+	//loop every minute
+	let ms = config.server_list.update_interval_secs * 1000;
+	let delay = ms - new Date() % ms;
+	setTimeout(this.updatePresence, delay);
 }
